requirements.txt
diffusers==0.24.0
einops==0.7.0
gradio==4.20.1
huggingface_hub==0.21.4
imageio==2.27.0
numpy==1.24.3
omegaconf==2.3.0
packaging==23.2
Pillow==10.1.0
rembg==2.0.55
safetensors==0.3.2
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1
transformers==4.27.0
trimesh==4.0.5
git+https://github.com/tatsy/torchmcubes.git
